feat: add --schools filter and improve grad year / OCR accuracy
- Add --schools flag to scan-resumes for school-based filtering
- Extend graduation year patterns to support '27年应届生' and '27年毕业'
- Add OCR text cleanup for common Vision misreadings ('gent/' -> 'agent/')
- Fix scrollList to handle fractional scrollTop values
- Fix OpenOnlineResume to scroll virtual lists to find candidates
- Update playbook docs

+++ boss/candidates.go
@@ -172,11 +172,11 @@ func scrollList(client *browser.Client, step int) (bool, error) {
 		return false, fmt.Errorf("scroll failed: %w", err)
 	}
 	var r struct {
-		BeforeTop int    `json:"beforeTop"`
-		AfterTop  int    `json:"afterTop"`
-		MaxScroll int    `json:"maxScroll"`
-		AtBottom  bool   `json:"atBottom"`
-		Error     string `json:"error,omitempty"`
+		BeforeTop float64 `json:"beforeTop"`
+		AfterTop  float64 `json:"afterTop"`
+		MaxScroll float64 `json:"maxScroll"`
+		AtBottom  bool    `json:"atBottom"`
+		Error     string  `json:"error,omitempty"`
 	}
 	if err := json.Unmarshal(raw, &r); err != nil {
 		return false, fmt.Errorf("parse scroll result failed: %w", err)
